refactor(dashboard): extract JSX for RepositoryInfoPanel component
