--- define-props-for-filtercontrols-PLAN.md
@@ -1,4 +1,0 @@
-# Define Props for FilterControls
-
-## Implementation Approach
-Review the FilterControls component to verify that the Props interface is correctly defined with all necessary state variables and functions (activityMode, dateRange, activeFilters, installations, loading, handleModeChange, handleDateRangeChange, handleOrganizationChange, session). Ensure proper typing for all props and proper function signatures.
