+++ extract-jsx-for-repositoryinfopanel-PLAN.md
@@ -1,0 +1,4 @@
+# Extract JSX for RepositoryInfoPanel
+
+## Implementation Approach
+Identify and extract the repository information panel JSX from the dashboard page, including the toggle button and repository list/summary, move it to the RepositoryInfoPanel component, and update the dashboard page to use the new component. Ensure all styles and functionality are preserved.
